fixed an exception with static {}
