+++ src/test/java/org/gwtproject/safecss/shared/GwtSafeStylesUtilsTest.java
@@ -15,7 +15,6 @@
  */
 package org.gwtproject.safecss.shared;
 
-import com.google.gwt.core.client.GWT;
 import com.google.gwt.junit.client.GWTTestCase;
 import org.gwtproject.dom.style.shared.Display;
 import org.gwtproject.dom.style.shared.TextAlign;
@@ -187,7 +186,7 @@ public class GwtSafeStylesUtilsTest extends GWTTestCase {
   }
 
   public void testFromTrustedNameAndValueInvalidName() {
-    if (GWT.isProdMode()) {
+    if (!"on".equals(System.getProperty("superdevmode"))) {
       // fromTrustedNameAndValue only catches errors in dev mode.
       return;
     }
@@ -210,7 +209,7 @@ public class GwtSafeStylesUtilsTest extends GWTTestCase {
   }
 
   public void testFromTrustedNameAndValueInvalidValue() {
-    if (GWT.isProdMode()) {
+    if (!"on".equals(System.getProperty("superdevmode"))) {
       // fromTrustedNameAndValue only catches errors in dev mode.
       return;
     }
@@ -233,7 +232,7 @@ public class GwtSafeStylesUtilsTest extends GWTTestCase {
   }
 
   public void testFromTrustedNameAndValueValidName() {
-    if (GWT.isProdMode()) {
+    if (!"on".equals(System.getProperty("superdevmode"))) {
       // fromTrustedNameAndValue only catches errors in dev mode.
       return;
     }
@@ -249,7 +248,7 @@ public class GwtSafeStylesUtilsTest extends GWTTestCase {
   }
 
   public void testFromTrustedNameAndValueValidValue() {
-    if (GWT.isProdMode()) {
+    if (!"on".equals(System.getProperty("superdevmode"))) {
       // fromTrustedNameAndValue only catches errors in dev mode.
       return;
     }
@@ -271,11 +270,4 @@ public class GwtSafeStylesUtilsTest extends GWTTestCase {
   private void assertEquals(String cssValue, SafeStyles safeStyles) {
     assertEquals(cssValue, safeStyles.asString());
   }
-
-//  private boolean isClientMode() {
-//    if (System.getProperty("superdevmode") == null) {
-//      return false;
-//    }
-//    return System.getProperty("superdevmode").equals("on") || System.getProperty("superdevmode").equals("off");
-//  }
 }
